entropy: widen the range-decoder window to 64 bits
Port dav1d's 64-bit coding window (src/msac.h ec_win = size_t, ctx_refill
at src/msac.c) onto the libaom-shaped reader: dif goes uint32 -> uint64 in
Reader and Cursor, and refillState buffers 6-7 bytes per refill through one
big-endian 8-byte load instead of 2-3 bytes, halving refill frequency. The
window width only changes WHEN bytes are buffered, never the symbol
arithmetic: every decode still reads and renormalizes the top 16 bits, the
subtract guards (coded >= lower, dif >= window) keep borrows out of the
buffered low bits, and goav1's XOR-into-ones window convention (libaom
entdec.c od_ec_dec_refill) is kept byte-for-byte.

The tell origin stays -14: pos*8 - cnt is window-width invariant (each
buffered byte advances both by 8), so libaom's 10 - (WINDOW - 8) formula,
which bakes in its fixed 32-bit window, must not be rescaled or every
BitsRead consumer (tile trailing-bit checks) would shift by 32.

The 32-bit arm64 kernels (readBitTrustedARM64, readCDF4HighTokenUpdateArch)
are tagged out and routed to their pure-Go cores; the 64-bit re-port
follows in the next commit. Verified against an embedded 32-bit reference
over randomized streams (symbols, rng, tell, window top-16 identical,
including past-end reads) and byte-exact on the strict extended
conformance suite (226/226).

+++ internal/av1/entropy/reader_bit_generic.go
@@ -2,8 +2,6 @@
 //
 // See LICENSE for the BSD-2-Clause grant and NOTICE for the AOM attribution.
 
-//go:build !arm64 || purego || goav1_trace_rng
-
 package entropy
 
 // ReadBitTrusted decodes one equiprobable bit from cursor state.
@@ -18,7 +16,7 @@ func (c *Cursor) ReadBitTrusted() uint8 {
 	cnt := int32(c.cnt)
 	split := (rangeValue >> 8) << 7
 	split += ecMinProb
-	window := split << (ecWindow - 16)
+	window := uint64(split) << (ecWindow - 16)
 
 	bit := uint8(1)
 	nextRange := split
