+++ internal/av1/entropy/trace_off.go
@@ -12,10 +12,10 @@ const (
 
 // traceCDFRead is a no-op stub when the goav1_trace_rng build tag is not set.
 // Keeping the helper inlineable preserves the entropy reader hot path.
-func traceCDFRead(_ uint16, _ int, _ uint32, _ uint32, _ int) {}
+func traceCDFRead(_ uint16, _ int, _ uint64, _ uint32, _ int) {}
 
 // traceBoolRead is a no-op stub when the goav1_trace_rng build tag is not set.
-func traceBoolRead(_ uint16, _ uint32, _ uint32, _ int) {}
+func traceBoolRead(_ uint16, _ uint64, _ uint32, _ int) {}
 
 // TraceLabel is a no-op stub when the goav1_trace_rng build tag is not set.
 func TraceLabel(_ string, _ ...any) {}
